비즈니스 로직으로 이동
